Added maps in case we need them
Speed Maps are maps which are relatively empty (i.e. not filled with a lot of barrels), which forces players to rush to the middle to fight it out with each other near the treasure chest.

Pirate Maps are maps which are relatively full (i.e. filled with a lot of barrels), which forces players to strategize and optimize their route to the treasure chest.